+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
-## [0.0.3]
+## [Unreleased]
+
+### Added
+- display project name in the toolbar.
+- added a "Refresh" button to update the panel content.
+- added an "Analyze Project" button to run a static analyzer on the project.
+- add the logic for displaying unused packages and files.
+- show files count and lines of code count.
+- improved styling for buttons and results display.
+- a remove button to remove package.
+- add button that will run `Dart: Add Dependency` command.
+
+### Changed
+- move the open panel button from pubspec.yaml to the toolbar.
+
+### Fixed
+- Fix #1
+- '^' will not be removed from the version constraint.
+
+## [0.0.4]
 
 - First release of Pub Manager.